Moved interface out and added comments.
